requirements.txt
amqp==5.0.9
aniso8601==7.0.0
anyjson==0.3.3
apturl==0.5.2
asgiref==3.6.0
attrs==21.2.0
autopep8==2.0.1
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
celery==5.2.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.5
django-admin-interface==0.24.2
django-admin-sortable2==2.1.4
django-ckeditor==6.0.0
django-colorfield==0.8.0
django-constance==2.8.0
django-cors-headers==3.13.0
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-hvad==1.8.0
django-js-asset==1.2.2
django-parler==2.3
django-parler-rest==2.2
django-picklefield==3.0.1
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
ephem==4.1.3
fasteners==0.14.1
future==0.18.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.3.0
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-memcached==1.59
pytz==2022.7
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
Rx==1.6.3
SecretStorage==3.3.1
singledispatch==4.0.0
six==1.15.0
sqlparse==0.4.3
systemd-python==234
text-unidecode==1.3
texttable==1.6.4
tomli==2.0.1
ttystatus==0.38
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
